ported to 64bit time for ITCH

+++ ITCH/include/time.hpp
@@ -2,6 +2,6 @@
 
 namespace ITCH41 {
 
-     uint32_t nanoseconds_since_midnight();
+     uint64_t nanoseconds_since_midnight();
 
 }
