+++ ITCH/time.cpp
@@ -6,14 +6,14 @@
 
 namespace ITCH41 {
 
-    uint32_t nanoseconds_since_midnight() {
+    uint64_t nanoseconds_since_midnight() {
         using namespace std::chrono;
 
         auto now = system_clock::now();
         auto today = floor<days>(now);
         auto time_since_midnight = now - today;
 
-        return static_cast<uint32_t>(duration_cast<nanoseconds>(time_since_midnight).count());
+        return static_cast<uint64_t>(duration_cast<nanoseconds>(time_since_midnight).count());
     }
 
 }
